fix(core): stop the main barrel from re-exporting filesystem implementations
@gitgov/core is published to npm and its package.json promises a runtime-agnostic
root entrypoint, with Node-only code behind ./fs. Three barrels broke that promise
and dragged node:fs into the published bundle.

EARS-CI02 caught it once the guardrail was re-armed. It had been it.skip'd for
seven months, and its report was dead on top of that: the catch called fail(), a
Jasmine global Jest does not expose, so it raised ReferenceError before printing.
It could detect without being able to say what it found.

  key_provider/index.ts   dropped `export * from './fs'` and './memory'
  file_lister/index.ts    same. github and prisma already complied
  agent_discovery         split: the two pure functions stay, discoverInstalledAgents
                          moves to fs/fs_agent_discovery.ts and its three tests
                          move with it, keeping ids DISC-A1..A3

The notes left in each barrel list every implementation of that module and why
they are not re-exported, following record_store/index.ts.

The defect was real but latent: every consumer that runs is Node, and saas-web —
the only browser consumer — imports seven symbols from @gitgov/core and all seven
are `import type`, so no value import reaches the client bundle. Only static
analysis of the built bundle sees it, and that analysis was switched off.

tsc caught every break: six KeyProvider.FsKeyProvider call sites in the CLI and two
source_auditor tests importing FsFileLister from the barrel.

Guardrail down from five violations to four. The rest come from llm/cli and
diagram_generator.
