+++ packages/core/src/agent_discovery/fs/index.ts
@@ -1,0 +1,1 @@
+export { discoverInstalledAgents } from './fs_agent_discovery';
